pyproject.toml
[tool.poetry]
name = "discord-server-bridge"
version = "0.1.0"
description = ""
authors = ["Michael <m.n.stewart@live.co.uk>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^1.7.3"
python-decouple = "^3.8"
flask = "^2.3.3"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-mock = "*"
coverage = "^7.2.2"
pre-commit = "2.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
